requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.3
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.7.0
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-extensions==2.1.6
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docopt==0.6.2
entrypoints==0.3
facebook-sdk==3.1.0
flake8==3.7.7
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
jwt==0.6.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
pep8==1.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyjwt==1.7.1
pylint==2.3.1
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
